sign-file,extract-cert: move common SSL helper functions to a header
commit 300e6d4116f956b035281ec94297dc4dc8d4e1d3 upstream.

Couple error handling helpers are repeated in both tools, so
move them to a common header.

Tested-by: R Nageswara Sastry <rnsastry@linux.ibm.com>
